+++ src/link/file.ts
@@ -2,7 +2,7 @@ import {generateForChild, generateForSelf, fixByRemove, fixByInsert} from '../..
 import {BoundingRect} from '../../lib/rect';
 import Parser from '../../index';
 import {LinkBaseToken} from './base';
-import {ImageParameterToken} from '../imageParameter';
+import {ImageParameterToken, extensions} from '../imageParameter';
 import type {
 	TokenTypes,
 	Config,
@@ -40,8 +40,7 @@ const frame = new Map([
 	argTypes = new Set<TokenTypes>(['arg']),
 	transclusion = new Set<TokenTypes>(['template', 'magic-word']),
 	horizAlign = new Set(['left', 'right', 'center', 'none']),
-	vertAlign = new Set(['baseline', 'sub', 'super', 'top', 'text-top', 'middle', 'bottom', 'text-bottom']),
-	extensions = new Set(['tiff', 'tif', 'png', 'gif', 'jpg', 'jpeg', 'webp', 'xcf', 'pdf', 'svg', 'djvu']);
+	vertAlign = new Set(['baseline', 'sub', 'super', 'top', 'text-top', 'middle', 'bottom', 'text-bottom']);
 
 /**
  * a more sophisticated string-explode function
@@ -447,7 +446,7 @@ export abstract class FileToken extends LinkBaseToken {
 	getFrame(): string | Title | undefined {
 		const [arg] = this.getFrameArgs(),
 			val = arg?.name;
-		return val === 'manualthumb' ? this.normalizeTitle(arg!.getValue() as string, 6, {page: ''}) : val;
+		return val === 'manualthumb' ? arg!.thumb : val;
 	}
 
 	/**